Make file write sync operation to be async and rearrange return statements
Make file write sync operation to be async because file write operation can be of blocking in nature and rearrange return statements in order to use event driven concepts of node js. 

Async call of file write means, we need to pass a third arg which would be an argument to a function (call back function) and will be executed once the file write operation is done. The same concept can be used for error handling later on. 

Return statement needs to be applied on the end event, because otherwise nodejs will just register a new end event and run that independently while running the code which is outside the end event -> which is to return the response with status code 302. But that is wrong, because 302 should be returned only after file write is done. And also, code (the 'My first page' code) should not  be run after end event because it doesnt make sense.