Handling different routes using middlewares
handling "/" and "/add-product" routes. As "/" means anything that starts with "/", this middleware is to be put at last. 

Also for handling routes, if we are in the right middle ware, we dont call next(), instead just send the response. 

The first dummy middleware shows that all URL Paths are going to match with it and it is going to run for all the calls but as it calls next(), the execution continues.