Create shop.js for storing routes of the online shop - like homepage
It is a good idea to decouple our code. Hence the route logic is moved to /routes folder. This specific class would store logic of routing for the online shop like homepage. 

Express.js provides framework to create routes. We first need to create router using express js and then instead of calling "app.use", replace with router.use()

Here in the code, we use router.get() to clearly define that this is a GET API. 
Major difference between use() and get()/post() is that the latter perform the exact match of the URL Paths. 

In normal cases with use(), "/" would correspond to everything that starts with / , but with. get(), it has to be. exact match and a GET call.