Use the already created admin and shop routers
Thi would refactor the routing logic. 

the routes which are exported by admin and shop.js, they are treated as actual middlewares only and hence can be easily configured with app.use(). 

And now because we are using router.get/post() instead of router.use(), hence exact matching of URL paths would be performed. Hence the exact order of app.use(routes) in the app,js file does not matter , but is still important from code quality point of view.