Add 404 Error Page and filter paths for /admin routes
As part of error handling and user experience, we need a 404 error page. As this is a middleware that needs to be executed when none of the other middle wares get a match, hence it should be placed at the bottom. 

The path is not provided for this , as it will default to a generic one, which can accept all URLs. 

We also added filtering of paths.
Instead of user reaching out to /add-product, we want users to go to /admin/add-products (GET and POST), meaning we are expecting all the APIs of admin route to start with /admin, hence we use pass that path in app.use(). The [PATH] variable , which is the first param of app.use(), can accept multiple paths and hence this works out of the box.